:sparkles: add bubble menu to text editors
