+++ app/admin/grant-verdict.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import { Button } from '@/components/button'
-import { TextEditor, useTextEditor } from '@/components/editor'
+import { TextEditor } from '@/components/editor'
+import { useTextEditor } from '@/utils/use-text-editor'
 import { Row } from '@/components/layout/row'
 import { Modal } from '@/components/modal'
 import { HorizontalRadioGroup } from '@/components/radio-group'
