+++ app/create/create-project-form.tsx
@@ -6,7 +6,8 @@ import { Input } from '@/components/input'
 import { Button } from '@/components/button'
 import { useRouter } from 'next/navigation'
 import { TOTAL_SHARES } from '@/db/project'
-import { TextEditor, useTextEditor } from '@/components/editor'
+import { TextEditor } from '@/components/editor'
+import { useTextEditor } from '@/utils/use-text-editor'
 import Link from 'next/link'
 import { add, format, isAfter, isBefore } from 'date-fns'
 import { Col } from '@/components/layout/col'
