+++ app/projects/[slug]/comments.tsx
@@ -3,7 +3,8 @@ import { Profile } from '@/db/profile'
 import { CommentAndProfile, sendComment } from '@/db/comment'
 import { UserAvatarAndBadge } from '@/components/user-link'
 import { formatDistanceToNow } from 'date-fns'
-import { RichContent, TextEditor, useTextEditor } from '@/components/editor'
+import { RichContent, TextEditor } from '@/components/editor'
+import { useTextEditor } from '@/utils/use-text-editor'
 import { Project } from '@/db/project'
 import { ArrowUturnRightIcon } from '@heroicons/react/24/outline'
 import { PaperAirplaneIcon } from '@heroicons/react/24/solid'
