+++ app/projects/[slug]/edit-description.tsx
@@ -2,7 +2,8 @@
 
 import { Button, IconButton } from '@/components/button'
 import { useSupabase } from '@/db/supabase-provider'
-import { useTextEditor, TextEditor } from '@/components/editor'
+import { TextEditor } from '@/components/editor'
+import { useTextEditor } from '@/utils/use-text-editor'
 import { Project } from '@/db/project'
 import { useState } from 'react'
 import { PencilIcon } from '@heroicons/react/24/outline'
