+++ app/rounds/[roundSlug]/edit-round.tsx
@@ -2,7 +2,8 @@
 
 import { Button } from '@/components/button'
 import { useSupabase } from '@/db/supabase-provider'
-import { useTextEditor, TextEditor } from '@/components/editor'
+import { TextEditor } from '@/components/editor'
+import { useTextEditor } from '@/utils/use-text-editor'
 import { useState } from 'react'
 import { useRouter } from 'next/navigation'
 import { Round } from '@/db/round'
