+++ components/editor.tsx
@@ -1,47 +1,10 @@
 'use client'
-import Link from '@tiptap/extension-link'
-import {
-  useEditor,
-  EditorContent,
-  Editor,
-  mergeAttributes,
-} from '@tiptap/react'
+import { EditorContent, Editor, BubbleMenu } from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
 import clsx from 'clsx'
 import { DisplayMention } from './user-mention/mention-extension'
-import { linkClass } from './site-link'
 import { generateReact } from './tiptap-utils'
-import Placeholder from '@tiptap/extension-placeholder'
-
-export function useTextEditor(
-  content?: any,
-  className?: string,
-  placeholder?: string
-) {
-  const editor = useEditor({
-    editorProps: {
-      attributes: {
-        class: clsx(
-          proseClass('md'),
-          'py-[.5em] px-4 h-full bg-white border border-gray-300 rounded-md min-h-[5em] focus:outline-orange-500',
-          className
-        ),
-      },
-    },
-    extensions: [
-      StarterKit,
-      DisplayLink,
-      DisplayMention,
-      Placeholder.configure({
-        placeholder,
-        emptyEditorClass:
-          'before:content-[attr(data-placeholder)] before:text-gray-500 before:float-left before:h-0 cursor-text',
-      }),
-    ],
-    content: content ?? '',
-  })
-  return editor
-}
+import { DisplayLink } from '@/utils/use-text-editor'
 
 export function TextEditor(props: {
   editor: Editor | null
@@ -50,25 +13,47 @@ export function TextEditor(props: {
   const { editor, children } = props
   return (
     <div className="relative w-full rounded-lg shadow-sm transition-colors">
+      {editor && (
+        <BubbleMenu
+          editor={editor}
+          tippyOptions={{ duration: 100 }}
+          className="flex items-center rounded bg-white py-2 shadow"
+        >
+          <button
+            onClick={() => editor.chain().focus().toggleBold().run()}
+            className={clsx(
+              'border-r border-gray-300 px-2 font-bold',
+              editor.isActive('bold') ? 'is-active' : ''
+            )}
+          >
+            B
+          </button>
+          <button
+            onClick={() => editor.chain().focus().toggleItalic().run()}
+            className={clsx(
+              'border-r border-gray-300 px-2 italic',
+              editor.isActive('bold') ? 'is-active' : ''
+            )}
+          >
+            I
+          </button>
+          <button
+            onClick={() => editor.chain().focus().toggleStrike().run()}
+            className={clsx(
+              'border-r border-gray-300 px-2 line-through',
+              editor.isActive('bold') ? 'is-active' : ''
+            )}
+          >
+            S
+          </button>
+          <span className="px-2 text-sm font-thin">Ctrl + V for hyperlink</span>
+        </BubbleMenu>
+      )}
       <EditorContent editor={editor} className="w-full" />
       {children}
     </div>
   )
 }
-
-// From Manifold's editor
-const proseClass = (size: 'sm' | 'md' | 'lg') =>
-  clsx(
-    'prose max-w-none leading-relaxed',
-    'prose-a:text-orange-600 prose-a:no-underline',
-    size === 'sm' ? 'prose-sm' : 'text-md',
-    size === 'sm' && 'prose-p:my-0 prose-ul:my-0 prose-ol:my-0 prose-li:my-0',
-    size === 'md' && 'prose-p:my-2 prose-ul:my-2 prose-ol:my-2 prose-li:my-2',
-    '[&>p]:prose-li:my-0',
-    'text-gray-900 prose-blockquote:text-gray-600',
-    'prose-a:font-light prose-blockquote:font-light font-light',
-    'break-anywhere'
-  )
 
 // TODO: Extract to server component
 export function RichContent(props: {
@@ -85,6 +70,10 @@ export function RichContent(props: {
     DisplayMention,
   ])
 
+  function proseClass(size: string): import('clsx').ClassValue {
+    throw new Error('Function not implemented.')
+  }
+
   return (
     <div
       className={clsx(
@@ -98,10 +87,3 @@ export function RichContent(props: {
     </div>
   )
 }
-
-export const DisplayLink = Link.extend({
-  renderHTML({ HTMLAttributes }) {
-    delete HTMLAttributes.class // Only use our classes (don't duplicate on paste)
-    return ['a', mergeAttributes(HTMLAttributes, { class: linkClass }), 0]
-  },
-})
