+++ package.json
@@ -26,6 +26,7 @@
     "@supabase/supabase-js": "^2.26.0",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/line-clamp": "^0.4.2",
+    "@tiptap/extension-bubble-menu": "^2.0.4",
     "@tiptap/extension-link": "^2.0.0-beta.218",
     "@tiptap/extension-mention": "^2.0.2",
     "@tiptap/extension-placeholder": "^2.0.3",
