+++ pages/api/comment-notifications.ts
@@ -8,7 +8,7 @@ import { calculateFullTrades } from '@/utils/math'
 import { getTxnsByProject } from '@/db/txn'
 import StarterKit from '@tiptap/starter-kit'
 import { DisplayMention } from '@/components/user-mention/mention-extension'
-import { DisplayLink } from '@/components/editor'
+import { DisplayLink } from '@/utils/use-text-editor'
 import { parseMentions } from '@/utils/parse'
 import { Comment } from '@/db/comment'
 import { JSONContent } from '@tiptap/core'
