+++ utils/use-text-editor.ts
@@ -1,0 +1,63 @@
+'use client'
+import Link from '@tiptap/extension-link'
+import { mergeAttributes, useEditor } from '@tiptap/react'
+import StarterKit from '@tiptap/starter-kit'
+import clsx from 'clsx'
+import { DisplayMention } from '@/components/user-mention/mention-extension'
+import { linkClass } from '@/components/site-link'
+import Placeholder from '@tiptap/extension-placeholder'
+import { BubbleMenu } from '@tiptap/extension-bubble-menu'
+
+export function useTextEditor(
+  content?: any,
+  className?: string,
+  placeholder?: string
+) {
+  const editor = useEditor({
+    editorProps: {
+      attributes: {
+        class: clsx(
+          proseClass('md'),
+          'py-[.5em] px-4 h-full bg-white border border-gray-300 rounded-md min-h-[5em] focus:outline-orange-500',
+          className
+        ),
+      },
+    },
+    extensions: [
+      StarterKit,
+      DisplayLink,
+      DisplayMention,
+      Placeholder.configure({
+        placeholder,
+        emptyEditorClass:
+          'before:content-[attr(data-placeholder)] before:text-gray-500 before:float-left before:h-0 cursor-text',
+      }),
+      BubbleMenu.configure({
+        element: document.querySelector('.menu') as HTMLElement,
+      }),
+    ],
+    content: content ?? '',
+  })
+  return editor
+}
+
+// From Manifold's editor
+const proseClass = (size: 'sm' | 'md' | 'lg') =>
+  clsx(
+    'prose max-w-none leading-relaxed',
+    'prose-a:text-orange-600 prose-a:no-underline',
+    size === 'sm' ? 'prose-sm' : 'text-md',
+    size === 'sm' && 'prose-p:my-0 prose-ul:my-0 prose-ol:my-0 prose-li:my-0',
+    size === 'md' && 'prose-p:my-2 prose-ul:my-2 prose-ol:my-2 prose-li:my-2',
+    '[&>p]:prose-li:my-0',
+    'text-gray-900 prose-blockquote:text-gray-600',
+    'prose-a:font-light prose-blockquote:font-light font-light',
+    'break-anywhere'
+  )
+
+export const DisplayLink = Link.extend({
+  renderHTML({ HTMLAttributes }) {
+    delete HTMLAttributes.class // Only use our classes (don't duplicate on paste)
+    return ['a', mergeAttributes(HTMLAttributes, { class: linkClass }), 0]
+  },
+})
